Get rid of `abundanceControl()` in the species parameter controls. The abundance is not a species parameter and if we want such a control it should go onto the abundance tab.
Avoid duplicate updating of params object.

Make the undo functionality more robust.

Introduce `tuneParams_update_params()`. Closes #52
This should be called when the params object is updated and it counts the number of changes since the last steady in the `changes` attribute.

We are now including the call to `params()` in the call to `tuneParams_add_to_logs()`.